+++ tests/testthat/test-cjs_add_bars.R
@@ -7,8 +7,6 @@ test_that("cjs_add_bars works with source data", {
     chartjs() %>%
     cjs_add_bars(x = letters, y = numbers, label = 'from source data')
 
-  expect_identical(p$x$data$labels, sort(data$letters))
-
   p_data <-
     with(p$x$data$datasets[[1]],{
       data.frame(x = sapply(data, function(d) d$x),
@@ -25,8 +23,6 @@ test_that("cjs_add_bars works with source data - vectors", {
   p <-
     chartjs() %>%
     cjs_add_bars(x = LETTERS[c(1,3,2,4,5)], y = 5:1, label = 'from vectors')
-
-  expect_identical(p$x$data$labels, sort(LETTERS[c(1,3,2,4,5)]))
 
   p_data <-
     with(p$x$data$datasets[[1]],{
@@ -45,8 +41,6 @@ test_that("cjs_add_bars works with source data - vectors", {
     chartjs() %>%
     cjs_add_bars(x = x1, y = y1, label = 'from given data',
                  data = data.frame(x1 = LETTERS[c(5,2,4,3,1)], y1 = c(3,2,4,1,5)))
-
-  expect_identical(p$x$data$labels, sort(LETTERS[c(5,2,4,3,1)]))
 
   p_data <-
     with(p$x$data$datasets[[1]],{